Made generative capabilities (hopefully) reflect new structure.
