+++ EventStream/transformer/conditionally_independent_model.py
@@ -5,8 +5,8 @@ import torch
 from ..data.types import DataModality, PytorchBatch
 from .config import StructuredEventProcessingMode, StructuredTransformerConfig
 from .generation.generation_utils import StructuredGenerationMixin
-from .generative_model_base import GenerativeOutputLayerBase
 from .model_output import (
+    GenerativeOutputLayerBase,
     GenerativeSequenceModelLabels,
     GenerativeSequenceModelLosses,
     GenerativeSequenceModelOutput,
